+++ e2e/pages/login_page.py
@@ -3,8 +3,8 @@ from .base_page import BasePage
 class LoginPage(BasePage):
     def __init__(self, page):
         super().__init__(page)
-        # self.url = "http://localhost:8000/sign-in/"
-        self.url = "/sign-in/?next=/dashboard/"
+        # self.url = "/sign-in/?next=/dashboard/"
+        self.url = "/sign-in/"
         self.username_field = self.page.locator("#id_username_or_email")
         self.password_field = self.page.locator("#id_password")
         self.login_button = self.page.locator("#sign-in-button")
